analyzer/requirements.txt
nltk
textblob
matplotlib
seaborn
wordcloud
pandas
python-docx
scikit-learn
sumy
transformers 